refactor all the fetching services into their own directory & template
